feat: require Scheduler 6.4+ (#65)
* Rebase on current 1.x

* Update Symfony dependencies to version 6.4

* Remove backward compatibility with Symfony/Scheduler 6.3

The code removed was ensuring backward compatibility with the Symfony/Scheduler 6.3. As we are now using version 6.4, these checks and the associated methods are redundant and have been removed from `TaskInfo.php` and MessengerMonitorController.php.

* Update GitHub workflow to test only with Symfony 6.4

The changes are made to the GitHub workflow to only test with Symfony 6.4. The backward compatibility testing with Symfony 6.3 has been discontinued as it is no longer needed.