feat: Add message type filter in History (#78)
In the History module, a filter for message types has been introduced. This improvement includes the creation of a new method in Storage to retrieve distinct message types, allowing users to filter messages specifically by type. Additionally, a visual drop-down list was updated in the history.html.twig template to display and select the message types.